chore: move facility schemas to v2 and update refs
